[CALCITE-1243] Ensure signed fields are used in protobuf for specifying rows to return
It was found that the protobuf definitions where specifying some fields in some
messages as unsigned integers where Avatica was expecting negative values
to be meaningful. We failed to notice this in Avatica because the Java bindings
for protobuf are equivalent for signed and unsigned typed. This breaks clients
in other languages though.

Some confusing logic surrounding `maxRowCount` being used for two different
meanings (both maximum number of rows for the statement and the maximum number
of rows returned in the first frame) was also corrected since the values in
the protocol buffers were being changed.

The incorrect fields were deprecated, documented, and backwards-compatibally modified
so that older clients can continue to operate without change. Clients in languages
which do have differentiation between signedness in types will have to use the
newer version of the client and server to get this fix.

+++ avatica/server/src/main/java/org/apache/calcite/avatica/jdbc/JdbcResultSet.java
@@ -65,7 +65,7 @@ class JdbcResultSet extends Meta.MetaResultSet {
    * negative value (typically -1) returns an unlimited number of rows
    * in frames of the default frame size. */
   public static JdbcResultSet create(String connectionId, int statementId,
-      ResultSet resultSet, long maxRowCount) {
+      ResultSet resultSet, int maxRowCount) {
     try {
       Meta.Signature sig = JdbcMeta.signature(resultSet.getMetaData());
       return create(connectionId, statementId, resultSet, maxRowCount, sig);
@@ -75,7 +75,7 @@ class JdbcResultSet extends Meta.MetaResultSet {
   }
 
   public static JdbcResultSet create(String connectionId, int statementId,
-      ResultSet resultSet, long maxRowCount, Meta.Signature signature) {
+      ResultSet resultSet, int maxRowCount, Meta.Signature signature) {
     try {
       final Calendar calendar = Calendar.getInstance(DateTimeUtils.GMT_ZONE);
       final int fetchRowCount;
@@ -86,7 +86,7 @@ class JdbcResultSet extends Meta.MetaResultSet {
       } else if (maxRowCount > AvaticaStatement.DEFAULT_FETCH_SIZE) {
         fetchRowCount = AvaticaStatement.DEFAULT_FETCH_SIZE;
       } else {
-        fetchRowCount = (int) maxRowCount;
+        fetchRowCount = maxRowCount;
       }
       final Meta.Frame firstFrame = frame(null, resultSet, 0, fetchRowCount, calendar);
       if (firstFrame.done) {
